login redirect changes

+++ admin/delete.php
@@ -1,12 +1,16 @@
 <?php
 $id = $_GET['del'];
 
-$con = new mysqli('localhost', 'root', '','toystore');
+require '../dbconn.php';
+if (!isset($_SESSION['adminemail'])) {
+	$_SESSION['error'] = 'Login to continue';
+	header('location: /admin/');
+}
 $sql = "delete from toys where t_id='$id'";
-if ($con->query($sql) === TRUE) { ?>
+if ($conn->query($sql) === TRUE) { ?>
   <script type="text/javascript">location.href = 'manage.php';</script>
 <?php } else {
-  echo "Error deleting record: " . $con->error;
+  echo "Error deleting record: " . $conn->error;
 }
-$con->close();
+$conn->close();
 ?>
